refactored to include next server
