+++ src/apollo/services/ExonService.test.ts
@@ -1,4 +1,4 @@
-import {addTranscript, annotationEditorCommand} from './ProteinCodingService'
+import {addTranscript} from './ProteinCodingService'
 import {
   addOrganismWithDirectory,
   deleteOrganism,
@@ -12,6 +12,7 @@ import {User} from '../domain/User'
 import {Role} from '../domain/Role'
 import {sleep} from '../functions/Timing'
 import {GenomeAnnotationGroup} from '../domain/GenomeAnnotationGroup'
+import {annotationEditorCommand} from "./ApolloAPIService";
 
 const TEST_USER = 'test@test.com'
 const TEST_ORGANISM = 'testAnimal'
